Fix: 게시물 작성기능 에러수정
