+++ nextapp/src/components/modules/Post/EditPost.tsx
@@ -18,9 +18,10 @@ import codeSyntaxHighlight from '@toast-ui/editor-plugin-code-syntax-highlight';
 import { _axios } from "hooks/axios";
 import SubmitBtn from "components/ui/button/SubmitBtn";
 import axios from "axios";
+import { PostType } from "types/interface";
 import { CategoryData } from "types/types";
 
-const CreatePosts = styled.section`
+const EditPosts = styled.section`
 	display: flex;
 	flex-direction: column;
 	align-items: center;
@@ -59,16 +60,21 @@ const Label = styled.label`
 	margin-right: 10px;
 `;
 
-export default function CreatePost ({children}: {children: React.ReactNode}) {
+export default function EditPost ({ children }: {children: React.ReactNode}) {
+	const obj = window.sessionStorage.getItem('post-storage');
+	const data: PostType = JSON.parse(obj as string).state.nowPost;
+
 	const [ optionList, setOptionList] = useState<React.ReactElement[]>([
 		<option key='init' value='선택'>
 			선택
 		</option>
 	]);
 	const editorRef = useRef<any>(null);
-	const [ title, setTitle] = useState<string>('');
-	const [ categoryId, setCategoryId] = useState<string>('');
-	const [ contents, setContents] = useState<any>();
+	const [ title, setTitle] = useState<string>(data.title);
+	const [ categoryId, setCategoryId] = useState<string | number>(data.categoryId);
+	const [ contents, setContents] = useState<any>(data.contents);
+
+	// console.log(data.contents);
 
 	useEffect(() => {
 		const fetchCategory = async () => {
@@ -80,7 +86,9 @@ export default function CreatePost ({children}: {children: React.ReactNode}) {
 					</option>
 				)
 			});
-			setOptionList([ ...optionList, ...optionEl]);
+			const initData = categoryData.filter(el=> el.id === data.categoryId)[0];
+			const initCategory = <option key='init' value={initData.id}>{initData.title}</option>
+			setOptionList([ initCategory,...optionEl]);
 		}
 		fetchCategory();
 	// eslint-disable-next-line react-hooks/exhaustive-deps
@@ -96,12 +104,13 @@ export default function CreatePost ({children}: {children: React.ReactNode}) {
 			alert('항목을 모두 작성해주세요'); 
 		} else if (title && categoryId && contents){
 			try {
-				const req = await _axios.post('/post/write',{title, categoryId, contents})
+				const req = await axios.patch(`${process.env.NEXT_PUBLIC_CORS_URL}/post/update`,{ id: data.id, title, categoryId, contents})
+				console.log(req);
 				if(req){
 					window.location.replace(`${process.env.NEXT_PUBLIC_REDIRECT}`)
-					return alert('새로운 게시물이 등록되었습니다')
+					return alert('게시물이 수정되었습니다')
 				} else if(!req){
-					alert('게시물 등록 실패')
+					alert('게시물 수정 실패')
 				}
 			} catch (err) {
 				alert(err);
@@ -109,11 +118,10 @@ export default function CreatePost ({children}: {children: React.ReactNode}) {
 		}
 	};
 
-
 	return(
-		<CreatePosts>
+		<EditPosts>
 			<H1>
-				글 작성하기
+				글 수정하기
 			</H1>
 			<EditorBox>
 				<InfoBox>
@@ -121,7 +129,7 @@ export default function CreatePost ({children}: {children: React.ReactNode}) {
 					<input 
 						style={{borderRadius: '5px', padding:'5px', border:'1px solid gray', fontSize:'0.7rem', opacity:'0.5'}} 
 						type="text" 
-						placeholder="제목을 입력해주세요"
+						defaultValue={data.title}
 						onChange={(e)=>{setTitle(e.target.value)}}
 					/>
 				</InfoBox>
@@ -138,7 +146,7 @@ export default function CreatePost ({children}: {children: React.ReactNode}) {
 					ref={editorRef}
 					height="100%"
 					initialEditType='markdown' 
-					initialValue='type here!!!'	
+					initialValue={data.contents}	
 					previewStyle="vertical"
 					hideModeSwitch={true}
 					plugins={[[codeSyntaxHighlight, { highlighter: Prism }], colorSyntax ]}
@@ -165,6 +173,6 @@ export default function CreatePost ({children}: {children: React.ReactNode}) {
 				></Editor>
 				<SubmitBtn handler={submitHandler}/>
 			</EditorBox>
-		</CreatePosts>
+		</EditPosts>
 	)
 }
